Move fix scripts from test/ to scripts/
