+++ package.json
@@ -40,7 +40,7 @@
   "scripts": {
     "confluence": "node ./node_modules/mdn-confluence/main/generate.es6.js --output-dir=. --bcd-module=./index.js",
     "lint": "node test/lint",
-    "fix": "node test/fix",
+    "fix": "node scripts/fix",
     "stats": "node scripts/statistics",
     "release-notes": "node scripts/release-notes",
     "render": "node scripts/render",
