+++ scripts/fix-browser-order.js
@@ -47,4 +47,52 @@ const fixBrowserOrder = (filename) => {
   }
 }
 
+if (require.main === module) {
+  /**
+   * @param {string[]} files
+   */
+  function load(...files) {
+    for (let file of files) {
+      if (file.indexOf(__dirname) !== 0) {
+        file = path.resolve(__dirname, '..', file);
+      }
+
+      if (!fs.existsSync(file)) {
+        continue; // Ignore non-existent files
+      }
+
+      if (fs.statSync(file).isFile()) {
+        if (path.extname(file) === '.json') {
+          fixBrowserOrder(file);
+        }
+
+        continue;
+      }
+
+      const subFiles = fs.readdirSync(file).map((subfile) => {
+        return path.join(file, subfile);
+      });
+
+      load(...subFiles);
+    }
+  }
+
+  if (process.argv[2]) {
+    load(process.argv[2]);
+  } else {
+    load(
+      'api',
+      'css',
+      'html',
+      'http',
+      'svg',
+      'javascript',
+      'mathml',
+      'test',
+      'webdriver',
+      'webextensions'
+    );
+  }
+}
+
 module.exports = fixBrowserOrder;
